+++ src/plot_example.cpp
@@ -13,7 +13,7 @@ int main() {
     std::vector<double> x = {1, 2, 3, 4, 5};
     std::vector<double> y = {1, 4, 9, 16, 25};
 
-    plt.plot(x, y);
+    plt.plot({x, y});
     plt.show();
 
     return 0;