+++ src/sca_example.cpp
@@ -11,16 +11,16 @@ int main() {
 
     std::vector<long> figsize = {10, 10};
 
-    auto [fig, axes] = plt.subplots(1, 2, figsize);
+    auto [fig, axes] = plt.subplots({1, 2, figsize});
 
     plt.sca(axes[0]);
-    axes[0].plot({1, 2, 3, 4, 5}, {1, 4, 9, 16, 25});
+    axes[0].plot({{1, 2, 3, 4, 5}, {1, 4, 9, 16, 25}});
     axes[0].set_title("Line");
     axes[0].grid(true);
 
 
     plt.sca(axes[1]);
-    axes[1].plot({1, 2, 3, 4, 5}, {1, 4, 9, 16, 25});
+    axes[1].plot({{1, 2, 3, 4, 5}, {1, 4, 9, 16, 25}});
     axes[1].set_title("Bar");
     axes[1].grid(true);